+++ src/config-interface.ts
@@ -2,6 +2,22 @@ import * as vscode from 'vscode';
 
 export const sectionName = 'filePropertiesViewer';
 
+export const propertyRows = [
+	'name',
+	'directory',
+	'fullPath',
+	'realPath',
+	'size',
+	'created',
+	'changed',
+	'modified',
+	'accessed',
+	'permissions',
+	'mediaType',
+] as const;
+
+export type PropertyRowType = typeof propertyRows[number];
+
 export class Config
 {
 	static get section(): vscode.WorkspaceConfiguration & TypedConfig
@@ -17,6 +33,7 @@ interface TypedConfig
 	get(item: 'queryMediaInfo'): boolean;
 	get(item: 'outputStylePath'): string | null;
 	get(item: 'disableRelativeTimestamps'): boolean;
+	get(item: 'propertyRows'): string[];
 }
 
 export type SizeMode = 'kilo' | 'kibi';