+++ src/test/index.ts
@@ -16,7 +16,7 @@ async function go()
 	catch (e)
 	{
 		console.error('Tests failed.', e);
-		process.exit(0);
+		process.exit(1);
 	}
 }