+++ vite.config.ts
@@ -1,34 +1,41 @@
 import { defineConfig } from 'vite';
 
-export default defineConfig({
-	build: {
-		lib: {
-			entry: {
-				'extension': 'src/extension.ts',
-				'extension.test': 'src/test/extension.test.ts',
-			},
-			name: 'extension',
-			formats: ['cjs'],
+export default defineConfig((config) =>
+{
+	return {
+		test: {
+			environment: 'node',
+			include: ['src/test/**/*.unit.test.ts'],
 		},
-		minify: true,
-		rollupOptions: {
-			output: {
-				dir: 'out',
+		build: {
+			lib: {
+				entry: {
+					'extension': 'src/extension.ts',
+					'extension.test': 'src/test/extension.test.ts',
+				},
+				name: 'extension',
+				formats: ['cjs'],
 			},
-			external: [
-				'vscode',
-				'fs',
-				'path',
-				'os',
-				'util',
-				'child_process',
-				'events',
-				'timers',
-				'stream',
-				'buffer',
-				'assert',
-				'mocha',
-			],
+			minify: true,
+			rollupOptions: {
+				output: {
+					dir: 'out',
+				},
+					external: [
+						'vscode',
+						'fs',
+						'path',
+						'url',
+						'os',
+						'util',
+						'child_process',
+						'events',
+						'timers',
+						'stream',
+						'buffer',
+						'assert',
+					],
+			},
 		},
-	},
+	};
 });